+++ miden-vm/src/lib.rs
@@ -13,13 +13,14 @@ pub use miden_core::proof::{ExecutionProof, HashFunction};
 #[cfg(not(target_arch = "wasm32"))]
 pub use miden_processor::execute_sync;
 pub use miden_processor::{
-    DefaultHost, ExecutionError, ExecutionOptions, FastProcessor, FutureMaybeSend, Host, Kernel,
-    Program, ProgramInfo, StackInputs, SyncHost, ZERO, advice, crypto, execute, field,
-    operation::Operation, serde, trace, trace::ExecutionTrace, utils,
+    DefaultHost, ExecutionError, ExecutionOptions, ExecutionOutput, FastProcessor, FutureMaybeSend,
+    Host, Kernel, Program, ProgramInfo, StackInputs, SyncHost, TraceGenerationContext, ZERO,
+    advice, crypto, execute, field, operation::Operation, serde, trace, trace::ExecutionTrace,
+    utils,
 };
-#[cfg(not(target_arch = "wasm32"))]
-pub use miden_prover::prove_sync;
 pub use miden_prover::{InputError, ProvingOptions, StackOutputs, Word, prove};
+#[cfg(not(target_arch = "wasm32"))]
+pub use miden_prover::{prove_from_trace_sync, prove_sync};
 pub use miden_verifier::VerificationError;
 
 // (private) exports